+++ mobilesynth/mobilesynth/Classes/synth/controller.h
@@ -100,7 +100,7 @@ namespace synth {
         float GetSample();
         
         void GetFloatSamples(float* buffer, int size);
-        void GetInt32Sapmles(long* buffer, int size);
+        void GetInt32Sapmles(int* buffer, int size);
         void GetCharSamples(char *buffer, int size);
         
         void setFormat(QAudioFormat * f);
